make it better

+++ migrations/tables/friend_graphics__ages.js
@@ -12,15 +12,15 @@ TableFriendGraphicsAges.up = () => {
     migration.create('friend_graphics__ages', (table) => {
         table.integer('id').unsigned().increment();
         table.string('date_update');
-        table.integer('unknown').default(0);
-        table.integer('from0to14').default(0);
-        table.integer('from15to19').default(0);
-        table.integer('from20to24').default(0);
-        table.integer('from25to29').default(0);
-        table.integer('from30to34').default(0);
-        table.integer('from35to39').default(0);
-        table.integer('from40to44').default(0);
-        table.integer('from45to49').default(0);
-        table.integer('from50').default(0);
+        table.float('unknown').default(0);
+        table.float('from0to14').default(0);
+        table.float('from15to19').default(0);
+        table.float('from20to24').default(0);
+        table.float('from25to29').default(0);
+        table.float('from30to34').default(0);
+        table.float('from35to39').default(0);
+        table.float('from40to44').default(0);
+        table.float('from45to49').default(0);
+        table.float('from50').default(0);
     });
 };
