+++ migrations/tables/friend_graphics__genders.js
@@ -12,8 +12,8 @@ TableFriendGraphicsGenders.up = () => {
     migration.create('friend_graphics__genders', (table) => {
         table.integer('id').unsigned().increment();
         table.string('date_update');
-        table.integer('unknown').default(0);
-        table.integer('male').default(0);
-        table.integer('female').default(0);
+        table.float('unknown').default(0);
+        table.float('male').default(0);
+        table.float('female').default(0);
     });
 };
